Change of path due to refactoring in #6
